requirements.txt
werkzeug
pillow
keras
numpy
flask
tensorflow
pandas
opencv-python
